This is the best way i guess
