+++ core/src/com/sgparrish/woods/entity/Player.java
@@ -1,51 +1,9 @@
 package com.sgparrish.woods.entity;
 
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Input;
-import com.badlogic.gdx.math.Vector2;
-import com.sgparrish.woods.util.DebugRenderer;
-
-public class Player implements Entity {
-
-    public World world;
-    public final Vector2 position;
-    public final Vector2 velocity;
+public class Player extends PhysicsEntity {
 
     public Player() {
-        position = new Vector2();
-        velocity = new Vector2();
+        super();
     }
 
-    @Override
-    public void update(float delta) {
-        float vel = 0.005f;
-        if (Gdx.input.isKeyPressed(Input.Keys.DPAD_LEFT)) {
-            velocity.x -= vel;
-        } else if (Gdx.input.isKeyPressed(Input.Keys.DPAD_RIGHT)) {
-            velocity.x += vel;
-        } else {
-            // player.velocity.x = 0;
-        }
-        if (Gdx.input.isKeyPressed(Input.Keys.DPAD_UP)) {
-            velocity.y += vel;
-        } else if (Gdx.input.isKeyPressed(Input.Keys.DPAD_DOWN)) {
-            velocity.y -= vel;
-        } else {
-            // player.velocity.y = 0;
-        }
-
-        velocity.y -= 0.0001;
-
-        position.add(velocity);
-
-        if (world.getFromVector(position) != null) {
-            position.y = (int) (position.y + 1);
-            velocity.y = 0;
-        }
-    }
-
-    @Override
-    public void render() {
-        DebugRenderer.getInstance().renderPlayer(this);
-    }
 }
